split implementation of I2cController into separate files
